Move ParserM, ParseState into their own file
